account_backend/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
dj-database-url==3.0.1
Django==6.0
django-bcrypt==0.9.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
minio==7.2.20
pillow==12.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
py-bcrypt==0.4
pycparser==2.23
pycryptodome==3.23.0
PyJWT==2.10.1
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.2
